Add user authorisation actions checking, update tests

+++ src/main/java/hexlet/code/service/UserService.java
@@ -3,10 +3,12 @@ package hexlet.code.service;
 import hexlet.code.dto.users.UserCreateDTO;
 import hexlet.code.dto.users.UserDTO;
 import hexlet.code.dto.users.UserUpdateDTO;
+import hexlet.code.exception.ActionForbiddenException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.UserMapper;
 import hexlet.code.model.User;
 import hexlet.code.repository.UserRepository;
+import hexlet.code.util.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,6 +25,9 @@ public class UserService {
     @Autowired
     private UserMapper userMapper;
 
+    @Autowired
+    private UserUtils userUtils;
+
     public List<UserDTO> getAll() {
         var users = userRepository.findAll();
         return users.stream()
@@ -38,25 +43,36 @@ public class UserService {
 
     public UserDTO findByEmail(String email) {
         var user = userRepository.findByEmail(email)
-                    .orElseThrow(() -> new ResourceNotFoundException("User not found: " + email));
+                    .orElseThrow(() -> new ResourceNotFoundException("User not found"));
         return userMapper.map(user);
     }
 
     public UserDTO findById(Long id) {
         var user = userRepository.findById(id)
-                    .orElseThrow(() -> new ResourceNotFoundException("User not found: " + id));
+                    .orElseThrow(() -> new ResourceNotFoundException("User not found"));
         return userMapper.map(user);
     }
 
     public UserDTO update(@Valid UserUpdateDTO userData, Long id) {
+
         var user = userRepository.findById(id)
-                    .orElseThrow(() -> new ResourceNotFoundException("User not found: " + id));
+                    .orElseThrow(() -> new ResourceNotFoundException("User not found"));
+
+        var currentUser = userUtils.getCurrentUser();
+        if (null == currentUser || id != currentUser.getId()) {
+            throw new ActionForbiddenException("Update forbidden");
+        }
+
         userMapper.update(userData, user);
         userRepository.save(user);
         return userMapper.map(user);
     }
 
     public void delete(Long id) {
+        var currentUser = userUtils.getCurrentUser();
+        if (null == currentUser || id != currentUser.getId()) {
+            throw new ActionForbiddenException("Delete forbidden");
+        }
         userRepository.deleteById(id);
     }
 }
